requirements.txt
flask
flask_sqlalchemy
pyjwt
pytest
pytest-cov